build-notification-handler/requirements.txt
google-cloud-logging==3.9.0